% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-countryinfo.R
\docType{data}
\name{countryinfo}
\alias{countryinfo}
\title{countryinfo}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 304 rows and 4 columns.
}
\usage{
countryinfo
}
\description{
World Bank Data on countries
}
\keyword{datasets}
