% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-index-data-summarise.R
\name{.output_meta}
\alias{.output_meta}
\title{Write meta data file for user to modify}
\usage{
.output_meta(df)
}
\arguments{
\item{df}{dataframe in tidyindexR format}
}
\value{
Returns a summary of the raw data.
}
\description{
This analyses raw index data and summarises the data
}
\examples{
data(povstats)
index_create_meta(povstats)

}
\author{
David Hammond
}
\keyword{summarise}
