% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-calculate.R
\name{index_calculate}
\alias{index_calculate}
\title{Calculates Index}
\usage{
index_calculate(df)
}
\arguments{
\item{df}{dataframe in tidyindexR format}
}
\value{
Returns preprocessed data.
}
\description{
This prepares the raw data for use in the index by
}
\examples{
#NOT RUN BECAUSE NEED USER GENERATED META DATA FILE
# data(povstats)
# index_create_meta(df)
# df <- index_data_preprocess(povstats)
# df <- index_calculate(df)

}
\author{
David Hammond
}
\keyword{summarise}
