% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-create-corpus.R
\name{index_create_data_corpus}
\alias{index_create_data_corpus}
\title{Prepares raw data}
\usage{
index_create_data_corpus(df)
}
\arguments{
\item{df}{dataframe in tidyindexR format}

\item{index_meta_data_path}{where the meta data is stored}
}
\value{
Returns preprocessed data.
}
\description{
This prepares the raw data for use in the index by
}
\examples{
#NOT RUN BECAUSE NEED USER GENERATED META DATA FILE
# data(povstats)
# df <- index_data_preprocess(povstats)

}
\author{
David Hammond
}
\keyword{summarise}
