% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-index_meta_create.R
\name{index_create_meta}
\alias{index_create_meta}
\title{Create meta data table for user input}
\usage{
index_create_meta(df)
}
\arguments{
\item{df}{dataframe in tidyindexR format}
}
\value{
Returns a summary of the raw data.
}
\description{
This analyses raw index data and summarises the data
}
\examples{

#data(povstats)
#index_create_meta(povstats)

}
\author{
David Hammond
}
\keyword{summarise}
