% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-band.R
\name{index_data_band}
\alias{index_data_band}
\title{Prepares bands data between 0 and 1}
\usage{
index_data_band(values, worst_value = min(values), best_value = max(values))
}
\arguments{
\item{values}{values to band}

\item{worst_value}{value that will be zero. Everything 'lower' than this will also be given a score of zero.}

\item{best_value}{value to be scaled to one. Everything 'higher' than this will also be given a score of one.}
}
\value{
Returns banded data.
}
\description{
This returns a 0 to 1 banding based on user inputs of weights and indicator direction
}
\examples{

}
\author{
David Hammond
}
