% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-pad.R
\name{index_data_pad}
\alias{index_data_pad}
\title{Pad and Interpolate}
\usage{
index_data_pad(df)
}
\arguments{
\item{df}{dataframe in tidyindexR format}
}
\value{
Returns a summary of the raw data.
}
\description{
This analyses raw index data and summarises the data
}
\author{
David Hammond
}
\keyword{summarise}
