% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-preprocess.R
\name{index_data_preprocess}
\alias{index_data_preprocess}
\title{Prepares raw data}
\usage{
index_data_preprocess(df)
}
\arguments{
\item{df}{dataframe in tidyindexR format}
}
\value{
Returns preprocessed data.
}
\description{
This prepares the raw data for use in the index by
}
\examples{
data(povstats)
df <- index_data_preprocess(povstats)

}
\author{
David Hammond
}
\keyword{summarise}
