% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_prepare.R
\name{index_data_summarise}
\alias{index_data_summarise}
\title{Summarise Raw Data}
\usage{
index_data_summarise(df)
}
\arguments{
\item{df}{dataframe in tidyindexR format}
}
\value{
Returns a summary of the raw data.
}
\description{
This analyses raw index data and summarises the data
}
\examples{
data(povstats)
df <- index_data_summarise(povstats)

}
\author{
David Hammond
}
\keyword{summarise}
