% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-index-data-summarise.R
\name{outliers}
\alias{outliers}
\title{Calculate IQR for outliers}
\usage{
outliers(x)
}
\arguments{
\item{x}{vector of values}
}
\value{
Returns lower and upper IQR values
}
\description{
This calculates IQR range for indicators
}
\author{
David Hammond
}
\keyword{summarise}
