% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/povstats.R
\docType{data}
\name{povstats}
\alias{povstats}
\title{povstats}
\format{An object of class \code{data.frame} with 10255 rows and 4 columns.}
\usage{
povstats
}
\description{
World Bank Data on Poverty
}
\keyword{datasets}
