% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-knn.R
\name{index_data_knn}
\alias{index_data_knn}
\title{KNN Interpolate}
\usage{
index_data_knn(df)
}
\arguments{
\item{df}{dataframe in tidyindexR format}
}
\value{
Returns a summary of the raw data.
}
\description{
This analyses raw index data and summarises the data
}
\author{
David Hammond
}
\keyword{summarise}
